move:#165 목표 설정 라우팅/폴더 구조 변경
