+++ src/app/(client)/set-goal/components/step-weight.tsx
@@ -7,37 +7,37 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
-type SetGoalAgeStepProps = {
+type StepWeightProps = {
   userName: string;
   nextStep: (data: number) => void;
 };
 
 const formSchema = z.object({
-  age: FormSchema.ONLY_NUMBER_SCHEMA
+  weight: FormSchema.NUMBER_WITH_ONE_DECIMAL_SCHEMA
 });
 
 type FormValues = z.infer<typeof formSchema>;
 
-const SetGoalAgeStep = ({ userName, nextStep }: SetGoalAgeStepProps) => {
+const StepWeight = ({ userName, nextStep }: StepWeightProps) => {
   const form = useForm<FormValues>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      age: ''
+      weight: ''
     }
   });
 
   const onSubmit = (data: FormValues) => {
-    nextStep(Number(data.age));
+    nextStep(Number(data.weight));
   };
 
   return (
     <>
       <div>
         <Typography as="h3" variant="title2" className="mb-2">
-          {userName}님의 <br /> 나이를 알려주세요
+          {userName}님의 <br /> 몸무게를 알려주세요
         </Typography>
         <Typography as="span" variant="body2" className="text-gray-600">
-          더 정확하고 맞춤화된 피드백을 위해 필요해요
+          몸무게를 기준으로 맞춤 칼로리를 계산해요
         </Typography>
       </div>
       <div className="space-y-2 pt-2">
@@ -46,16 +46,17 @@ const SetGoalAgeStep = ({ userName, nextStep }: SetGoalAgeStepProps) => {
             <div className="space-y-[2.3rem]">
               <FormField
                 control={form.control}
-                name="age"
+                name="weight"
                 render={({ field }) => {
-                  const hasError = !!form.formState.errors.age;
+                  const hasError = !!form.formState.errors.weight;
+
                   return (
                     <FormItem>
-                      <FormLabel className="sr-only">나이</FormLabel>
+                      <FormLabel className="sr-only">몸무게</FormLabel>
                       <FormControl>
-                        <Input {...field} type="text" inputMode="numeric" className="mb-2" />
+                        <Input {...field} type="number" inputMode="numeric" measure="kg" className="mb-2" />
                       </FormControl>
-                      {!hasError && <FormDescription>숫자만 입력해 주세요</FormDescription>}
+                      {!hasError && <FormDescription>최대 소수점 1자리수까지 입력할 수 있어요</FormDescription>}
                       <FormMessage />
                     </FormItem>
                   );
@@ -65,7 +66,7 @@ const SetGoalAgeStep = ({ userName, nextStep }: SetGoalAgeStepProps) => {
             <Button
               type="submit"
               className="fixed bottom-6 left-1/2 w-[calc(100%-2.5rem)] -translate-x-1/2"
-              disabled={!form.watch('age')}
+              disabled={!form.watch('weight')}
             >
               다음
             </Button>
@@ -76,4 +77,4 @@ const SetGoalAgeStep = ({ userName, nextStep }: SetGoalAgeStepProps) => {
   );
 };
 
-export default SetGoalAgeStep;
+export default StepWeight;
